+++ src/index.ts
@@ -1,10 +1,18 @@
+import { pool as poolUnclean } from './pool/index.js';
+import type { Except } from './types/utilities.js';
+
 export { define } from './define/index.js';
 export { parent } from './service/index.js';
-export { pool } from './pool/index.js';
 export { ConcurrencyOption } from './types/pool.js';
 export { Messenger, messengers } from './messenger/index.js';
 export { ServiceCluster } from './service_cluster/index.js';
+export { SharedMap } from './shared_map/index.js';
+export { Bytes } from './types/shared_map.js';
 
 export type { TaskDefinitions } from './types/definitions.js';
 export type { Nanolith } from './types/nanolith.js';
 export type { TaskWorkerOptions as LaunchTaskOptions, ServiceWorkerOptions as LaunchServiceOptions } from './types/workers.js';
+export type { SharedMapTransferData as SharedMapTransfer } from './types/shared_map.js';
+export type { MessengerTransferData as MessengerTransfer } from './types/messenger.js';
+
+export const pool = poolUnclean as Except<typeof poolUnclean, '__enqueue'>;