+++ src/shared_map/constants.ts
@@ -1,0 +1,3 @@
+import { encodeValue } from './utilities.js';
+
+export const NULL_ENCODED = encodeValue(new TextEncoder(), null);