+++ CHANGELOG.md
@@ -6,14 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `SharedMap` implementation
+- New `Bytes` enum export to help users with calculating sizes for their `SharedMap`s
+- `notifyAll()` method to `ServiceCluster` to send messages to all services on the cluster with one function call.
+
 ### Changed
 
 - Add `timeoutSecs` parameter to `prepareWritableToPortStream()` to configure how many seconds to wait for the stream to be accepted before rejecting the promise.
 - `TaskWorkerOptions` and `ServiceWorkerOptions` types to be exported as `LaunchTaskOptions` and `LaunchServiceOptions` instead.
-
-### Added
-
-- `notifyAll()` method to `ServiceCluster` to send messages to all services on the cluster with one function call.
+- Main file to export types `SharedArrayPair` and `MessengerTransferData`.
+- `SharedArrayPair` type name to `SharedMapTransferData`.
+- `Messenger.transfer()` to be a getter instead of a function.
+- Removed `Data` word from exports of `SharedMapTransferData` and `MessengerTransferData`.
+- `pool.option` to be a static property.
+- Entirely removed the `__enqueue` method from the type for the main exported `pool`
 
 ## [0.2.4] - 2022-08-12